+++ src/ga/gosvoh/commands/Remove.java
@@ -11,7 +11,7 @@ import java.util.concurrent.ConcurrentSkipListMap;
  * @author Vokhmin Aleksey {@literal <}vohmina2011{@literal @}yandex.ru{@literal >}
  */
 public class Remove implements Command {
-    private ConcurrentSkipListMap<Integer, Universe> map;
+    private ConcurrentSkipListMap<Long, Universe> map;
 
     /**
      * Конструктор класса
@@ -23,7 +23,7 @@ public class Remove implements Command {
     @Override
     public String execute(String[] cmd) {
         try {
-            map.remove(Integer.parseInt(cmd[1]));
+            map.remove(Long.parseLong(cmd[1]));
             new SaveMap().execute(cmd);
             return ("Элемент успешно удалён!");
         } catch (NumberFormatException | ArrayIndexOutOfBoundsException e) {
